TIP-451: move BulkFileExporter in processor folder
