+++ src/Pim/Bundle/ConnectorBundle/Resources/config/processors.yml
@@ -10,6 +10,8 @@ parameters:
     pim_connector.processor.normalization.product.class:       Pim\Component\Connector\Processor\Normalization\ProductProcessor
     pim_connector.processor.normalization.variant_group.class: Pim\Component\Connector\Processor\Normalization\VariantGroupProcessor
 
+    pim_connector.processor.bulk_file_exporter.class: Pim\Component\Connector\Processor\BulkFileExporter
+
 services:
     # Dummy processor
     pim_connector.processor.dummy_item:
@@ -197,14 +199,14 @@ services:
             - '@pim_catalog.repository.attribute'
             - '@pim_catalog.builder.product'
             - '@akeneo_storage_utils.doctrine.object_detacher'
-            - '@pim_connector.writer.file.product.bulk_file_exporter'
+            - '@pim_connector.processor.bulk_file_exporter'
 
     pim_connector.processor.normalization.variant_group:
         class: '%pim_connector.processor.normalization.variant_group.class%'
         arguments:
             - '@pim_serializer.normalizer.group'
             - '@akeneo_storage_utils.doctrine.object_detacher'
-            - '@pim_connector.writer.file.product.bulk_file_exporter'
+            - '@pim_connector.processor.bulk_file_exporter'
             - '@pim_catalog.updater.variant_group'
 
     pim_connector.processor.normalization.family:
@@ -236,3 +238,11 @@ services:
         arguments:
             - '@pim_serializer.normalizer.association_type'
             - '@akeneo_storage_utils.doctrine.object_detacher'
+
+    # File exporter
+    pim_connector.processor.bulk_file_exporter:
+        class: '%pim_connector.processor.bulk_file_exporter.class%'
+        arguments:
+            - '@pim_connector.writer.file.file_exporter'
+            - '@pim_connector.writer.file.media_exporter_path_generator'
+            - ['pim_catalog_file', 'pim_catalog_image']
