+++ src/Pim/Component/Connector/Processor/Normalization/VariantGroupProcessor.php
@@ -11,7 +11,7 @@ use Akeneo\Component\Batch\Step\StepExecutionAwareInterface;
 use Akeneo\Component\StorageUtils\Detacher\ObjectDetacherInterface;
 use Akeneo\Component\StorageUtils\Updater\ObjectUpdaterInterface;
 use Pim\Component\Catalog\Model\GroupInterface;
-use Pim\Component\Connector\Writer\File\BulkFileExporter;
+use Pim\Component\Connector\Processor\BulkFileExporter;
 use Symfony\Component\Serializer\Normalizer\NormalizerInterface;
 
 /**
